refactor(linter/plugins): move settings-related code into separate file (#15401)
Pure refactor. Move this code into a separate file to keep it grouped together.
